+++ services/processor/processor/update_from_kitsu.py
@@ -2,139 +2,12 @@ from typing import TYPE_CHECKING, Any
 
 import ayon_api
 import gazu
-from nxtools import logging, slugify
+from nxtools import logging
 
 from . import utils
 
 if TYPE_CHECKING:
     from .processor import KitsuProcessor
-
-
-def _ensure_asset_type_folder(
-    project_name: str,
-    entity_type_id: str,
-    asset_type_name: str,
-    folders_by_kitsu_id: dict[str, dict],
-) -> str | None:
-    """Return the AYON folder id for an asset type, creating it if needed.
-
-    Args:
-        project_name: The name of the AYON project.
-        entity_type_id: The Kitsu entity type id.
-        asset_type_name: The name of the asset type.
-        folders_by_kitsu_id: A dictionary of AYON folders by Kitsu id.
-
-    Returns:
-        The AYON folder id for the asset type.
-    """
-    asset_type_folder = folders_by_kitsu_id.get(entity_type_id)
-    if asset_type_folder:
-        return asset_type_folder["id"]
-
-    assets_root = folders_by_kitsu_id.get("asset")
-    if not assets_root:
-        # Create the Assets folder if it doesn't exist
-        assets_root_id = ayon_api.create_folder(
-            project_name,
-            name="Assets",
-            data={"kitsuId": "asset"},
-        )
-        folders_by_kitsu_id["asset"] = {"id": assets_root_id}
-    else:
-        assets_root_id = assets_root["id"]
-
-    asset_type_folder_id = ayon_api.create_folder(
-        project_name,
-        name=slugify(asset_type_name, separator="_"),
-        label=asset_type_name,
-        parent_id=assets_root_id,
-        data={"kitsuId": entity_type_id},
-    )
-    folders_by_kitsu_id[entity_type_id] = {"id": asset_type_folder_id}
-    return asset_type_folder_id
-
-
-def move_folders_by_asset_type(
-    project_name: str, entities: list[dict[str, Any]]
-) -> None:
-    """Re-parent AYON asset folders when their Kitsu asset type differs.
-
-    Skip moves for folders that already have published products. Leave old
-    asset-type folders in place even if they become empty.
-
-    Args:
-        project_name: The name of the AYON project.
-        entities: List of kitsu entities to process.
-    """
-    entities_ids: set[str] = {"asset"}
-    for entity in entities:
-        if entity.get("id"):
-            entities_ids.add(entity["id"])
-        if entity.get("entity_type_id"):
-            entities_ids.add(entity["entity_type_id"])
-
-    folders_by_kitsu_id = utils.get_ayon_folders_by_kitsu_ids(
-        project_name, entities_ids
-    )
-
-    folders_to_move: list[tuple[dict[str, Any], dict, str]] = []
-    for entity in entities:
-        entity_type_id = entity["entity_type_id"]
-        asset_type_name = entity["asset_type_name"]
-
-        asset_folder = folders_by_kitsu_id.get(entity["id"])
-
-        desired_parent_id = _ensure_asset_type_folder(
-            project_name,
-            entity_type_id,
-            asset_type_name,
-            folders_by_kitsu_id,
-        )
-        if not desired_parent_id:
-            logging.warning(
-                f"Cannot move asset {entity.get('name')}: "
-                "failed to resolve asset type folder"
-            )
-            continue
-
-        if asset_folder.get("parentId") == desired_parent_id:
-            continue
-
-        folders_to_move.append((entity, asset_folder, desired_parent_id))
-
-    if not folders_to_move:
-        return
-
-    folder_ids_with_products = {
-        product["folderId"]
-        for product in ayon_api.get_products(
-            project_name,
-            folder_ids=[
-                asset_folder["id"] for _, asset_folder, _ in folders_to_move
-            ],
-            fields={"folderId"},
-        )
-    }
-
-    for entity, asset_folder, desired_parent_id in folders_to_move:
-        if asset_folder["id"] in folder_ids_with_products:
-            logging.warning(
-                f"Skipping move of asset '{entity.get('name')}' to "
-                f"'{entity.get('asset_type_name')}': "
-                "folder has published products"
-            )
-            continue
-
-        logging.info(
-            f"Moving asset '{entity.get('name')}' to asset type "
-            f"'{entity.get('asset_type_name')}'"
-        )
-        ayon_api.update_folder(
-            project_name,
-            asset_folder["id"],
-            parent_id=desired_parent_id,
-        )
-
 
 def update_project(parent: "KitsuProcessor", data: dict[str, str]):
     logging.info(f"update_project: {data}")
@@ -191,7 +64,7 @@ def create_or_update_asset(parent: "KitsuProcessor", data: dict[str, str]):
         entities=[entity],
     )
     # Re-parent after push so the folder is guaranteed to exist in AYON
-    move_folders_by_asset_type(project_name, [entity])
+    utils.move_folders_by_asset_type(project_name, [entity])
     return result
